add batch

+++ src/main/java/net/zousys/gba/batch/entity/JobDetailsRepository.java
@@ -2,5 +2,5 @@ package net.zousys.gba.batch.entity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobDetailsRepository extends JpaRepository<JobDetails, Integer> {
+public interface JobDetailsRepository extends JpaRepository<JobDetails, Long> {
 }
